content/svg/content/test/test_selectSubString2.xhtml

<!DOCTYPE html>
<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=398825
-->
<head>
  <title>Test for Bug 398825</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=398825">Mozilla Bug 398825</a>
<p id="display"></p>
<div id="content" style="display: none"></div>

<iframe id="svg" src="selectSubString-helper.svg"></iframe>

<pre id="test">
<script class="testbody" type="application/javascript">
SimpleTest.waitForExplicitFinish();

function runTests()
{
  var document = $("svg").contentWindow.document;
  var text = document.getElementById("text");

  ok(!text.selectSubString, "unexpected selectSubString method");

  SimpleTest.finish();
}

function run()
{
  SpecialPowers.pushPrefEnv({'set': [['svg.text.css-frames.enabled', false]]},
                            runTests);

}

window.addEventListener("load", run, false);
</script>
</pre>
</body>
</html>
